refactored Pledge Hw test into a 'tool', not anymore run as JUnit test by Maven.
