+++ pom.xml
@@ -25,7 +25,6 @@
       <groupId>junit</groupId>
       <artifactId>junit</artifactId>
       <version>4.13.1</version>
-      <scope>test</scope>
     </dependency>
 
     <dependency>
